+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### Skin Composer Version 52 ###
-* Updated TextraTypist Playground to use TextraTypist 0.6.0.
+* Added option to skip animation in TextraTypist Playground.
+* Updated TextraTypist Playground to use TextraTypist 0.6.2.
 
 ### Skin Composer Version 51 ###
 * Changed appearance and improved functionality of the color picker.